frontend/requirements.txt
Flask==3.0.0
Jinja2==3.1.2
boto3==1.40.25
python-dotenv==1.1.1